Fixed ut and changing finalVideo's length + improved accuracy

+++ lib/services/video_processing_data.dart
@@ -5,7 +5,7 @@ class VideoProcessingData {
   final Duration originalDuration;
   double normalizedTimeFraction;
   Duration expectedDuration;
-  int samplingRate;
+  double samplingRate;
 
   VideoProcessingData({this.video, this.originalDuration});
 }
